на три

+++ src/main/java/com/web_project/shop/service/CategoryService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface CategoryService {
     public List<CategoryModel> findAllCategories();
 
-    public CategoryModel findCategoryById(int id);
+    public CategoryModel findCategoryById(Long id);
 
     public CategoryModel addCategory(CategoryModel category);
 
     public CategoryModel updateCategory(CategoryModel category);
 
-    public void deleteCategory(int id);
+    public void deleteCategory(Long id);
 }
